requirements.txt
# requirements.txt - Dependencies

Flask==2.3.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0